+++ packages/terra-date-picker/src/terra-dev-site/doc/example/DatePickerIncomplete.jsx
@@ -1,0 +1,57 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Field from 'terra-form-field';
+import DatePicker from 'terra-date-picker';
+import classNames from 'classnames/bind';
+import styles from './DatePickerExampleCommon.module.scss';
+
+const cx = classNames.bind(styles);
+
+const propTypes = {
+  /**
+   * The current DatePicker date if selected. Use for the selected date message.
+   */
+  selectedDate: PropTypes.node,
+};
+
+const defaultProps = {
+  selectedDate: '',
+};
+
+class DatePickerIncompleteExample extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { date: this.props.selectedDate };
+    this.handleDateChange = this.handleDateChange.bind(this);
+  }
+
+  handleDateChange(event, date) {
+    this.setState({ date });
+  }
+
+  render() {
+    return (
+      <div>
+        <p>
+          Selected ISO Date:
+          <span className={cx('date-wrapper')}>{this.state.date}</span>
+        </p>
+        <Field label="Enter Date" htmlFor="default-incomplete">
+          <DatePicker
+            name="date-input"
+            id="default-incomplete"
+            onChange={this.handleDateChange}
+            required
+            isIncomplete
+            {...this.props}
+          />
+        </Field>
+      </div>
+    );
+  }
+}
+
+DatePickerIncompleteExample.propTypes = propTypes;
+DatePickerIncompleteExample.defaultProps = defaultProps;
+
+export default DatePickerIncompleteExample;