+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerIncomplete.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from '../../common/DateTimePickerExampleTemplate';
+
+const DateTimePickerExample = () => (
+  <DateTimePickerExampleTemplate
+    isIncomplete
+    required
+  />
+);
+
+export default DateTimePickerExample;