+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -6,6 +6,19 @@ Unreleased
 ### Breaking Changes
 * Changed 'NavigationPromptCheckpoint' to have the same logic as 'notification-dialog' changes
 
+### Changed
+* Missing aria-labels added in examples.
+
+1.15.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+1.14.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 1.13.0 - (October 3, 2019)
 ------------------
 ### Changed